Modificacion error delete

+++ src/main/java/com/paula/services/BookService.java
@@ -1,6 +1,7 @@
 package com.paula.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -26,6 +27,12 @@ public class BookService {
     }
     
     public void deleteBookById(Integer bookId) {
-        bookRepository.deleteById(bookId);
+        Optional<Book> book = bookRepository.findById(bookId);
+
+        if (book.isPresent()) {
+			bookRepository.deleteById(bookId);
+		} else {
+			throw new RuntimeException("¡Error! El libro con id " + bookId + " no está en la base de datos.");
+		}
     }
 }
